refactoring: início de modularização do código
